[METAL] Update metal runtime to directly store kernel map
This PR updates the metal runtime storage format to
directly store the kernel map. This will also enable
us to leverage metallib binary format which may not
be compatible with previous string split.

It changes the binary format of the metal module.
We also added a version to enable easier future
update.