[FFI][REFACTOR] Migrate the Save/Load JSON to the new reflection (#18188)
This PR migrates the Save/Load JSON to the new reflection based mechanism.

This is a breaking change that updates the the JSON format
to ffi/extra/serialization to handle the serialization,
see the json graph schema comment in ffi/extra/serialization.h
for the format, which roughly aligns with the old style.

After this change, we no longer need node/reflection and reflection vtable.
We can also phase out TVM_REGISTER_NODE and TVM_REGISTER_OBJECT to have a single
place that defines the reflection.