[FFI] Support Opaque PyObject (#18270)
* [FFI] Support Opaque PyObject

This PR adds support of Opaque PyObject.
When a type in python is not natively supported by ffi, it will
now be converted to an Opaque PyObject on the backend, such opaque
object will retain their lifecycle automatically and can still be
used by registering python callbacks or store in container
and return to the frontend.

* Round of grammar polishment